Move const folder into libs folder
