+++ src/lib/networks/index.ts
@@ -1,0 +1,2 @@
+export { getNetworkInfo } from "./getNetworkInfo";
+export { getTokenInfo } from "./getTokenInfo";
